Removed installEventBus and watchEventBus
