--- src/lib/watch/WatchEventBus.ts
@@ -1,27 +1,0 @@
-import EventBus from '../utils/EventBus';
-
-
-export enum WatchEvents {
-  Start = 'Start',
-  FilesChanged = 'FilesChanged',
-  Enable = 'Enable',
-  Disable = 'Disable',
-  Stop = 'Stop',
-  Stopped = 'Stopped',
-}
-
-export default class WatchEventBus extends EventBus {
-  static readonly Events = WatchEvents;
-
-  readonly Events = WatchEvents;
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  emit(event: WatchEvents, ...args: any[]): boolean {
-    return super.emit(event, ...args);
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  on(event: WatchEvents, listener: (...args: any[]) => void): this {
-    return super.on(event, listener);
-  }
-}
